Resolved Comments
