+++ src/index.ts
@@ -12,8 +12,6 @@ app.use(cors());
 app.use(express.json()); // Used to parse JSON bodies
 app.use(express.urlencoded()); // Parse URL-encoded bodies using query-string library
 
-
-
 app.listen(port, async () => {
   console.log("Server is running on port :", port);
   await connect();
