+++ src/routes/index.ts
@@ -1,26 +1,21 @@
 import { Express } from "express";
 import "express-yields";
-import swaggerUi from "swagger-ui-express";
 import expenses from "../features/expenses";
 import expensesList from "../features/expenses-list";
-import { healthCheck } from "../features/health-check";
+import { healthCheck } from "../features/health-check/health-check";
 import stats from "../features/stats";
 import {
   errorHandler,
   catchAllRequestsLastRouteHandler,
   checkJwt,
+  swagger,
 } from "./middlewares";
 import users from "../features/users";
-import { swaggerDocument, swaggerOptions } from "../swagger";
 
 export const routes = (app: Express) => {
   app.get(...healthCheck);
   app.use(...stats);
-  app.use(
-    "/docs",
-    swaggerUi.serve,
-    swaggerUi.setup(swaggerDocument, swaggerOptions)
-  );
+  app.use(...swagger);
   app.use(checkJwt, () => { });
   app.use(...users);
   app.use(...expenses);
